move some tests to protobuf
