+++ adapter/protobuf/protobuf_test.go
@@ -1,4 +1,4 @@
-package entity
+package protobuf
 
 import (
 	"bytes"
